reomve footballFixturesAdverts from commercialFeatures
